Fix time to FLOAT & .toFixed(5)

+++ database/init.sql
@@ -26,7 +26,7 @@ CREATE TABLE `files` (
   `filename` VARCHAR(255) NOT NULL,
   `path` VARCHAR(100) DEFAULT NULL,
   `uploaded_at` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-  `time` INT default NULL,
+  `time` FLOAT default NULL,
   `size` FLOAT default NULL,
   PRIMARY KEY (`id`),
   KEY `user_id` (`user_id`)
